# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AetherGrid-Sovereign API",
    "description": "Operational API platform for the AetherGrid-Sovereign cascade prediction system.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/events/ingest": {
      "post": {
        "tags": [
          "Ingestion & Snapshots"
        ],
        "summary": "Ingest Event",
        "description": "Ingests versioned sensor/events. Guaranteed idempotent.",
        "operationId": "ingest_event_api_v1_events_ingest_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Ingest Event Api V1 Events Ingest Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshots/materialize": {
      "post": {
        "tags": [
          "Ingestion & Snapshots"
        ],
        "summary": "Materialize Snapshot",
        "description": "Triggers materialization of a new graph snapshot from verified events.",
        "operationId": "materialize_snapshot_api_v1_snapshots_materialize_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "parent_hash",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Hash"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Materialize Snapshot Api V1 Snapshots Materialize Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestPayload": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id",
            "description": "Stable unique ID for idempotency"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "description": "E.g., sensor_reading, trigger_event",
            "default": "sensor_reading"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "Actual record attributes matching schemas"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "data"
        ],
        "title": "IngestPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}